move defaults to own file cause it was causing a circular dep and breaking the static automation test
