+++ src/default-config.js
@@ -1,0 +1,21 @@
+const defaults = {
+  open: true,
+  log: true,
+  mode: 'production',
+  webpackPlugins: [],
+  plugins: [],
+  baseURL: '/',
+  static: false,
+  src: 'docs/',
+  dst: '_ignite/',
+  index: 'index.md',
+  port: 8008,
+  title: 'Documentation',
+  codeStyle: 'foundation',
+  color: 'auto',
+  selectedColor: 'auto',
+  logo: 'logo.svg',
+  bulmaTheme: 'default'
+};
+
+export default defaults;
